fix(earthquake_history): 推計震度モードのPlaceholderを注意文言表示に置換
